- added the flexibility to add the menu entries (Actions) created by the instances of FlatCAMTool, in any position within the application menu's.
install method of FlatCAMTool accept now the following kwargs:

'pos' = it is the menu where we want the Action installed. By default,
if no 'pos' arg is provided, the tool will be install in the menutool
menu (as before)

'before' = it is the position within the previously selected menu where,
before it, we will install our new Action (menu entry) generated by the
FlatCAMTool install method. By default, if no 'before' is provided, the
action will be installed in the current last position in the menu.